literals (true/false, nothing, empty)
