+++ RDCore.SDK/Model/AST/Directives/ModuleDirectives.cs
@@ -8,12 +8,25 @@ using System.Collections.Immutable;
 namespace RDCore.SDK.Model.AST.Directives;
 
 /// <summary>
+/// A <c>BoundNode</c> representing a <c>VB_Attribute</c> directive.
+/// </summary>
+/// <param name="SemanticId">A semantic <c>Uri</c> uniquely identifying this specific node.</param>
+/// <param name="Location">The <c>Location</c> of the directive.</param>
+/// <param name="Name">The name of the attribute.</param>
+/// <param name="ValueExpression">An expression node that statically evaluates to the value of the attribute.</param>
+/// <param name="Binding">An optional qualifier used for binding the attribute to the member it belongs to.</param>
+public record class AttributeDirectiveNode(Uri SemanticId, SourceLocation Location, string Name, BoundNode ValueExpression, string? Binding = null)
+    : BoundDirective(SemanticId, Location, [ValueExpression]);
+
+/// <summary>
 /// A <c>BoundNode</c> representing an <c>Option</c> module directive.
 /// </summary>
 /// <param name="SemanticId">A semantic <c>Uri</c> uniquely identifying this specific node.</param>
 /// <param name="Location">The <c>Location</c> of the directive.</param>
 /// <param name="ModuleOption">The <c>ModuleOptions</c> value being configured.</param>
-public record class ModuleOptionDirectiveNode(Uri SemanticId, SourceLocation Location, ModuleOptions ModuleOption) : BoundDirective(SemanticId, Location) { }
+public record class ModuleOptionDirectiveNode(Uri SemanticId, SourceLocation Location, ModuleOptions ModuleOption)
+    : BoundDirective(SemanticId, Location, []);
+
 /// <summary>
 /// A <c>BoundNode</c> representing a <c>Def&lt;Type&gt;</c> module directive.
 /// </summary>
@@ -22,7 +35,7 @@ public record class ModuleOptionDirectiveNode(Uri SemanticId, SourceLocation Loc
 /// <param name="Token">The <c>DefType</c> token mapping to a specific <c>VBType</c> (per the semantics defined in MS-VBAL 5.2.2 Implicit Definition Directives).</param>
 /// <param name="Mappings">The prefixing scheme defined by this directive.</param>
 public record class TypeDefDirectiveNode(Uri SemanticId, SourceLocation Location, string Token, ImmutableArray<DefTypePrefixMapping> Mappings) 
-    : BoundDirective(SemanticId, Location) 
+    : BoundDirective(SemanticId, Location, []) 
 {
     public VBType GetVBType(bool is64bit) => Token switch
     {
